refactor: Adicionando os testes
