--- src/service/login.ts
@@ -1,22 +1,0 @@
-import axios, { AxiosResponse } from "axios";
-import { loginURL } from "../consts/service";
-import { LoginData } from "../models/service";
-
-interface LoginResponse {
-  status: number;
-}
-
-async function fetchLogin(loginData: LoginData): Promise<LoginResponse> {
-  try {
-    const response: AxiosResponse<LoginResponse> = await axios.post(
-      loginURL,
-      loginData
-    );
-    return response.data;
-  } catch (error) {
-    console.log(error);
-    throw error;
-  }
-}
-
-export default fetchLogin;
